Some reloc of functions and basic benchmarks
